TbitBrowscap - fast lightweight library for parsing php_browscap.ini and determining browser capabilities. Intended to replace the browscap library used in the .quark framework 20170613 - first version completed, tests performed succesfull on 7 different browsers
Only bitBrowscap.php file is required. The simplest way of using the library is to execute it as a singleton and obtain an object with browser capabilities as attributes:

$browser_info = TbitBrowscap::instance()->getBrowser();

If one whishes to obtain an array instead of an object:

$browser_info = TbitBrowscap::instance()->getBrowser(null, true);

The first argument specifies the user agent string (when null the $_SERVER['HTTP_USER_AGENT'] is used instead):

$browser_info = TbitBrowscap::instance()->getBrowser('some user agent string', true);

There are three attributes if one needs to change the class behaviour:

TbitBrowscap::instance()->cacheDir - where the cache tree is kept, default value is 'cache'
TbitBrowscap::instance()->iniFilename - the path to php_browscap.ini file. Default value is 'browscap.ini
TbitBrowscap::instance()->maxNodesPerCacheFile - the number of tree nodes stored per cache file in order to optimize loading times (it's a compromise between number of loading operations versus the size of each loading and parsing operation)